Fixed signed/unsigned comparisions in Transport

+++ cantera18/branches/scons/Cantera/src/transport/LiquidTransport.h
@@ -347,7 +347,7 @@ namespace Cantera {
      *
      *
      */
-     virtual void getSpeciesFluxes(int ndim, 
+     virtual void getSpeciesFluxes(size_t ndim,
 				  const doublereal* grad_T, 
 				  int ldx, const doublereal* grad_X, 
 				  int ldf, doublereal* fluxes);
@@ -361,7 +361,7 @@ namespace Cantera {
      *
      *
      */
-    virtual void getSpeciesFluxesExt(int ldf, doublereal* fluxes);
+    virtual void getSpeciesFluxesExt(size_t ldf, doublereal* fluxes);
 
 
     //! Solve the stefan_maxell equations for the diffusive fluxes.
